MyBlog237 v0.1.0 --server

+++ controllers/user_controller.js
@@ -3,6 +3,9 @@ import mongoose from "mongoose";
 
 export const getUsers = async (req, res) => {
   try {
+    if (req.userId && req.role !== process.env.ROLE) {
+      return res.status(401).json({ message: "ADMINISTRATOR CAPABILITIES" });
+    }
     const users = await User.find();
     res.status(200).json(users);
   } catch (err) {
@@ -16,6 +19,11 @@ export const deleteUser = async (req, res) => {
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(404).send("No user with that id");
   }
+
+  if (req.userId && req.role !== process.env.ROLE) {
+    return res.status(401).json({ message: "ADMINISTRATOR CAPABILITIES" });
+  }
+
   await User.findByIdAndRemove(id);
   res.json({ message: "user deleted" });
 };
